+++ ktlint-cli-reporter-baseline/src/main/kotlin/com/pinterest/ktlint/cli/reporter/baseline/BaselineReporter.kt
@@ -3,8 +3,13 @@ package com.pinterest.ktlint.cli.reporter.baseline
 import com.pinterest.ktlint.cli.reporter.core.api.KtlintCliError
 import com.pinterest.ktlint.cli.reporter.core.api.KtlintCliError.Status.FORMAT_IS_AUTOCORRECTED
 import com.pinterest.ktlint.cli.reporter.core.api.ReporterV2
+import java.io.File
 import java.io.PrintStream
+import java.nio.file.Path
+import java.nio.file.Paths
 import java.util.concurrent.ConcurrentHashMap
+import kotlin.io.path.pathString
+import kotlin.io.path.relativeToOrSelf
 
 public class BaselineReporter(
     private val out: PrintStream,
@@ -24,7 +29,10 @@ public class BaselineReporter(
         out.println("""<?xml version="1.0" encoding="utf-8"?>""")
         out.println("""<baseline version="1.0">""")
         for ((file, errList) in acc.entries.sortedBy { it.key }) {
-            out.println("""    <file name="${file.escapeXMLAttrValue()}">""")
+            // Store error in baseline always a relative path. This allows a baseline file to be stored inside a repository and after
+            // checking out this repository on a different path, the baseline will still be respected.
+            val relativeFile = Paths.get(file).relativeLocation()
+            out.println("""    <file name="${relativeFile.escapeXMLAttrValue()}">""")
             for ((line, col, ruleId, _) in errList) {
                 out.println("""        <error line="$line" column="$col" source="$ruleId" />""")
             }
@@ -39,4 +47,13 @@ public class BaselineReporter(
             .replace("'", "&apos;")
             .replace("<", "&lt;")
             .replace(">", "&gt;")
+
+    private fun Path.relativeLocation() =
+        relativeToOrSelf(ROOT_DIR_PATH)
+            .pathString
+            .replace(File.separatorChar, '/')
+
+    private companion object {
+        val ROOT_DIR_PATH: Path = Paths.get("").toAbsolutePath()
+    }
 }